[MOV] defaults: move helper to related file
