+++ src/plugins/core/default.ts
@@ -1,7 +1,7 @@
 import { CellPosition, Color, Dimension, Format, HeaderIndex, Style, UID, Zone } from "../..";
 import { DEFAULT_STYLE } from "../../constants";
 import { PositionMap } from "../../helpers/cells/position_map";
-import { getItemId, ItemsDic } from "../../helpers/data_normalization";
+import { ItemsDic, mapToId, mapToValue } from "../../helpers/data_normalization";
 import {
   deepCopy,
   deepEquals,
@@ -737,40 +737,6 @@ export class DefaultPlugin extends CorePlugin<defaultState> implements defaultSt
   // Export
   // ---------------------------------------------------------------------------
 
-  mapToId<T>(defaults: defaultValue<T>, dict: ItemsDic<T>): sparseDefaultValue<number> {
-    const defaultsIds: sparseDefaultValue<number> = {
-      colDefault: {},
-      rowDefault: {},
-    };
-    if (defaults.sheetDefault) {
-      defaultsIds.sheetDefault = getItemId(defaults.sheetDefault, dict);
-    }
-    for (const colIndex in defaults.colDefault) {
-      defaultsIds.colDefault![colIndex] = getItemId(defaults.colDefault[colIndex], dict);
-    }
-    for (const rowIndex in defaults.rowDefault) {
-      defaultsIds.rowDefault![rowIndex] = getItemId(defaults.rowDefault[rowIndex], dict);
-    }
-    return defaultsIds;
-  }
-
-  mapToValue<T>(defaultIds: sparseDefaultValue<number>, dict: ItemsDic<T>): defaultValue<T> {
-    const defaults: defaultValue<T> = {
-      colDefault: [],
-      rowDefault: [],
-    };
-    if (defaultIds.sheetDefault) {
-      defaults.sheetDefault = dict[defaultIds.sheetDefault];
-    }
-    for (const colIndex in defaultIds.colDefault) {
-      defaults.colDefault![colIndex] = dict[defaultIds.colDefault[colIndex]];
-    }
-    for (const rowIndex in defaultIds.rowDefault) {
-      defaults.rowDefault![rowIndex] = dict[defaultIds.rowDefault[rowIndex]];
-    }
-    return defaults;
-  }
-
   mapStyleToId(defaults: defaultStyle, dict: ItemsDic<Style>): sparseDefaultValue<number> {
     const colDefaultDict = defaultDict({});
     const rowDefaultDict = defaultDict({});
@@ -798,7 +764,7 @@ export class DefaultPlugin extends CorePlugin<defaultState> implements defaultSt
       rowDefault[row] = value;
     }
 
-    return this.mapToId(
+    return mapToId(
       {
         colDefault,
         rowDefault,
@@ -813,7 +779,7 @@ export class DefaultPlugin extends CorePlugin<defaultState> implements defaultSt
       const sheetFormat = this.format[sheet.id];
       sheet.defaultFormat =
         sheetFormat && !isObjectEmptyRecursive(sheetFormat)
-          ? this.mapToId(sheetFormat, data.formats)
+          ? mapToId(sheetFormat, data.formats)
           : undefined;
       const sheetStyle = this.style[sheet.id];
       sheet.defaultStyle =
@@ -828,10 +794,10 @@ export class DefaultPlugin extends CorePlugin<defaultState> implements defaultSt
       this.history.update(
         "format",
         sheet.id,
-        sheet.defaultFormat && this.mapToValue(sheet.defaultFormat, data.formats)
+        sheet.defaultFormat && mapToValue(sheet.defaultFormat, data.formats)
       );
       if (sheet.defaultStyle) {
-        const defaultStyle = this.mapToValue(sheet.defaultStyle, data.styles);
+        const defaultStyle = mapToValue(sheet.defaultStyle, data.styles);
         for (const key in defaultStyle.sheetDefault) {
           this.history.update(
             "style",
